+++ service-statistics-unsd/src/test/java/org/oskari/statistics/plugins/unsd/UnsdIndicatorValuesFetcherTest.java
@@ -1,7 +1,6 @@
-package fi.nls.oskari.control.statistics.plugins.unsd;
+package org.oskari.statistics.plugins.unsd;
 
 import fi.nls.oskari.control.statistics.data.IndicatorValue;
-import fi.nls.oskari.control.statistics.plugins.unsd.parser.UnsdIndicatorDataParserTest;
 import fi.nls.test.util.ResourceHelper;
 import org.json.JSONException;
 import org.junit.Test;
