+++ service-statistics-unsd/src/test/resources/org/oskari/statistics/plugins/unsd/IndicatorData2015.json
