+++ service-statistics-unsd/src/test/resources/org/oskari/statistics/plugins/unsd/Indicators.json
